Conflict solved, packaged moved to 'test' package
